+++ muna/beta/metadata/__init__.py
@@ -8,7 +8,6 @@ from ._torch import KVCacheConfig, TorchExporter
 from .coreml import CoreMLInferenceMetadata
 from .cuda import CudaArchitecture, GeneratedCudaKernelInferenceMetadata
 from .executorch import ExecuTorchInferenceBackend, ExecuTorchInferenceMetadata
-from .iree import IREEInferenceBackend, IREEInferenceMetadata
 from .litert import LiteRTInferenceMetadata
 from .llama import LlamaCppBackend, LlamaCppInferenceMetadata
 from .mlx import MLXInferenceMetadata, MLXInferenceSessionMetadata
@@ -17,10 +16,7 @@ from .openvino import OpenVINOInferenceMetadata
 from .qnn import QnnInferenceBackend, QnnInferenceMetadata, QnnInferenceQuantization
 from .sglang import SGLangInferenceMetadata
 from .tensorrt import TensorRTInferenceMetadata, TensorRTInferenceSessionMetadata
-from .tensorrt_llm import (
-    TensorRTLLMInferenceMetadata, TensorRTLLMQuantization,
-    TensorRTLLMKVCacheQuantization
-)
+from .tensorrt_llm import TensorRTLLMInferenceMetadata
 from .tensorrt_rtx import TensorRTRTXInferenceMetadata
 from .tflite import TFLiteInterpreterMetadata
 from .vllm import vLLMInferenceMetadata
@@ -29,7 +25,6 @@ CompileMetadata = (
     CoreMLInferenceMetadata                 |
     ExecuTorchInferenceMetadata             |
     #GeneratedCudaKernelInferenceMetadata    | # WIP
-    IREEInferenceMetadata                   |
     LiteRTInferenceMetadata                 |
     LlamaCppInferenceMetadata               |
     MLXInferenceMetadata                    |